LayoutTests/editing/undo-manager/undo-manager-delete-stale-undo-items-expected.txt

E
Verifies that JSUndoItems are deleted when they are no longer needed by the platform clipboard. This test requires WebKitTestRunner.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS objectCountBeforeClearingUndoStack - objectCountAfterClearingUndoStack is >= 580
PASS successfullyParsed is true

TEST COMPLETE
